refactor: 패키지 구조 변경, 리팩토링
- 응답 반복 코드 제거
- 응답을 HttpStatus enum을 사용하도록 수정
