+++ README.md
@@ -37,9 +37,9 @@
   - [x] 응답 URL을 index.html로 지정
   - [x] 회원가입시 입력한 모든 데이터를 추출해 User 객체를 생성한다.
 ### 리팩토링
-- [ ] 응답 반복 코드를 제거한다.
-  - [ ] 응답을 관리하는 클래스를 만든다.
-  - [ ] 응답 Http status를 ENUM으로 관리한다.
+- [x] 응답 반복 코드를 제거한다.
+  - [x] 응답을 관리하는 클래스를 만든다.
+  - [x] 응답 Http status를 enum으로 관리한다.
 #### 기능 요구사항 5.
 - [ ] 로그인 성공 실패 분기 처리를 한다.
   - [ ] 앞 단계에서 회원가입할 때 생성한 User 객체를 DataBase.addUser() 메서드를 활용해 RAM 메모리에 저장한다.
