--- src/main/java/webserver/header/exception/InvalidPathException.java
@@ -1,7 +1,0 @@
-package webserver.header.exception;
-
-public class InvalidPathException extends RuntimeException{
-    public InvalidPathException(String message) {
-        super(message);
-    }
-}
