--- src/main/java/webserver/header/exception/InvalidProtocolException.java
@@ -1,7 +1,0 @@
-package webserver.header.exception;
-
-public class InvalidProtocolException extends RuntimeException{
-    public InvalidProtocolException(String message) {
-        super(message);
-    }
-}
