--- src/main/java/webserver/header/exception/InvalidRequestParamsException.java
@@ -1,7 +1,0 @@
-package webserver.header.exception;
-
-public class InvalidRequestParamsException extends RuntimeException {
-    public InvalidRequestParamsException(String message) {
-        super(message);
-    }
-}
