--- src/main/java/webserver/header/exception/InvalidUriException.java
@@ -1,7 +1,0 @@
-package webserver.header.exception;
-
-public class InvalidUriException extends RuntimeException{
-    public InvalidUriException(String message) {
-        super(message);
-    }
-}
