--- src/main/java/webserver/header/exception/InvalidVersionException.java
@@ -1,7 +1,0 @@
-package webserver.header.exception;
-
-public class InvalidVersionException extends RuntimeException{
-    public InvalidVersionException(String message) {
-        super(message);
-    }
-}
