+++ src/main/java/webserver/request/header/RequestHeaderInfos.java
@@ -1,0 +1,35 @@
+package webserver.request.header;
+
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+
+public class RequestHeaderInfos {
+    private static final int REQUEST_LINE_NEXT_IDX = 1;
+    private static final String HEADER_DELIMITER = ": ";
+    private static final String HOST = "Host";
+    private static final String CONTENT_LENGTH = "Content-Length";
+    private static final String DEFAULT_LENGTH = "0";
+    private final Map<String, String> infos;
+
+    RequestHeaderInfos(String[] infos) {
+        this.infos = Collections.unmodifiableMap(createHeaderInfo(infos));
+    }
+
+    String host() {
+        return infos.get(HOST);
+    }
+
+    int contentsLength() {
+        return Integer.parseInt(infos.getOrDefault(CONTENT_LENGTH, DEFAULT_LENGTH));
+    }
+
+    private static Map<String, String> createHeaderInfo(String[] headerInfos) {
+        Map<String, String> result = new HashMap<>();
+        for (int i = REQUEST_LINE_NEXT_IDX; i < headerInfos.length; i++) {
+            String[] keyValue = headerInfos[i].split(HEADER_DELIMITER);
+            result.put(keyValue[0], keyValue[1]);
+        }
+        return result;
+    }
+}
