+++ src/main/java/webserver/request/header/exception/InvalidPathException.java
@@ -1,0 +1,7 @@
+package webserver.request.header.exception;
+
+public class InvalidPathException extends RuntimeException{
+    public InvalidPathException(String message) {
+        super(message);
+    }
+}
