+++ src/main/java/webserver/request/header/exception/InvalidProtocolException.java
@@ -1,0 +1,7 @@
+package webserver.request.header.exception;
+
+public class InvalidProtocolException extends RuntimeException{
+    public InvalidProtocolException(String message) {
+        super(message);
+    }
+}
