+++ src/main/java/webserver/request/header/exception/InvalidRequestParamsException.java
@@ -1,0 +1,7 @@
+package webserver.request.header.exception;
+
+public class InvalidRequestParamsException extends RuntimeException {
+    public InvalidRequestParamsException(String message) {
+        super(message);
+    }
+}
