+++ src/main/java/webserver/request/header/exception/InvalidUriException.java
@@ -1,0 +1,7 @@
+package webserver.request.header.exception;
+
+public class InvalidUriException extends RuntimeException{
+    public InvalidUriException(String message) {
+        super(message);
+    }
+}
