+++ src/main/java/webserver/request/header/exception/InvalidVersionException.java
@@ -1,0 +1,7 @@
+package webserver.request.header.exception;
+
+public class InvalidVersionException extends RuntimeException{
+    public InvalidVersionException(String message) {
+        super(message);
+    }
+}
