+++ src/main/java/webserver/service/RequestHandler.java
@@ -1,9 +1,11 @@
 package webserver.service;
 
 import java.io.BufferedReader;
+import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
 
@@ -11,9 +13,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import utils.IOUtils;
-import webserver.header.request.RequestHeader;
-import webserver.header.request.requestline.RequestLine;
+import webserver.request.header.RequestHeader;
 import webserver.method.HttpMethod;
+import webserver.response.get.GetIndexHtmlResponse;
 
 public class RequestHandler implements Runnable {
     private static final String NEXT_LINE = "\n";
@@ -29,11 +31,12 @@ public class RequestHandler implements Runnable {
         logger.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
                 connection.getPort());
 
-        try (InputStream in = connection.getInputStream()) {
+        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             BufferedReader br = new BufferedReader(new InputStreamReader(in, StandardCharsets.UTF_8));
             String line = br.readLine();
 
             StringBuilder requestContents = new StringBuilder(line);
+            DataOutputStream dos = new DataOutputStream(out);
 
             while (isEnd(line)) {
                 printlnLine(line);
@@ -43,18 +46,19 @@ public class RequestHandler implements Runnable {
             }
 
             RequestHeader header = RequestHeader.create(requestContents.toString());
-            RequestLine requestLine = header.requestLine();
 
-            if (isGet(requestLine)) {
-                ResponseGetHandler responseGetHandler = new ResponseGetHandler(connection);
-                responseGetHandler.run(requestLine);
+            if (isGet(header)) {
+                ResponseGetHandler responseGetHandler = new ResponseGetHandler();
+                GetIndexHtmlResponse response = new GetIndexHtmlResponse();
+                byte[] body = response.response(header.index());
+                writeAndFlush(responseGetHandler.handle(header, body.length), body, dos);
                 return;
             }
 
-            if (isPost(requestLine)) {
-                ResponsePostHandler responsePostHandler = new ResponsePostHandler(connection);
+            if (isPost(header)) {
+                ResponsePostHandler responsePostHandler = new ResponsePostHandler();
                 String requestBody = IOUtils.readData(br, header.contentLength());
-                responsePostHandler.run(header, requestLine, requestBody);
+                writeAndFlush(responsePostHandler.handle(header, requestBody), EMPTY.getBytes(), dos);
             }
         } catch (IOException e) {
             logger.error(e.getMessage());
@@ -69,11 +73,21 @@ public class RequestHandler implements Runnable {
         return !EMPTY.equals(line);
     }
 
-    private boolean isGet(RequestLine requestLine) {
+    private boolean isGet(RequestHeader requestLine) {
         return HttpMethod.isGet(requestLine.httpMethod());
     }
 
-    private boolean isPost(RequestLine requestLine) {
+    private boolean isPost(RequestHeader requestLine) {
         return HttpMethod.isPost(requestLine.httpMethod());
+    }
+
+    private static void writeAndFlush(String header, byte[] body, DataOutputStream dos) {
+        try {
+            dos.writeBytes(header);
+            dos.write(body, 0, body.length);
+            dos.flush();
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
     }
 }
