+++ src/main/java/webserver/service/exception/InvalidRequestException.java
@@ -1,0 +1,7 @@
+package webserver.service.exception;
+
+public class InvalidRequestException extends RuntimeException{
+    public InvalidRequestException(String message) {
+        super(message);
+    }
+}
