--- src/test/java/webserver/header/request/requestline/ProtocolTest.java
@@ -1,22 +1,0 @@
-package webserver.header.request.requestline;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import webserver.header.exception.InvalidProtocolException;
-
-class ProtocolTest {
-
-    @DisplayName("지원되지 않는 protocol의 경우 예외가 발생해야 한다.")
-    @Test
-    void createFailedByProtocolName() {
-        // given
-        // when
-        // then
-        assertThatThrownBy(() -> Protocol.from("TEST"))
-                .isInstanceOf(InvalidProtocolException.class)
-                .hasMessage("유효하지 않은 프로토콜 입니다.");
-    }
-}
